chore: isolate watcher tests in its domain
