+++ tests/watcher/files/mod.rs
@@ -1,0 +1,2 @@
+mod create;
+mod delete;
