+++ tests/watcher/follow/mod.rs
@@ -1,0 +1,2 @@
+mod friends;
+mod notification;
