+++ tests/watcher/mentions/mod.rs
@@ -1,0 +1,2 @@
+mod notification;
+mod raw;
