+++ tests/watcher/posts/reply.rs
@@ -1,4 +1,4 @@
-use super::utils::WatcherTest;
+use crate::watcher::utils::WatcherTest;
 use anyhow::Result;
 use pubky_common::crypto::Keypair;
 use pubky_nexus::{
