+++ tests/watcher/tags/mod.rs
@@ -1,0 +1,4 @@
+mod post;
+mod post_notification;
+mod user;
+mod user_notification;
