+++ tests/watcher/users/mod.rs
@@ -1,0 +1,1 @@
+mod raw;
